refactor: make register printing reusable
Code which converts KMP-spcific binary representation to native Python
types is useful on its own.
