+++ src/pykmp/tool/__main__.py
@@ -13,7 +13,7 @@ import attrs
 import click
 
 import pykmp
-from pykmp import client, codec, constants, messages
+from pykmp import client, constants, messages, registers
 
 if TYPE_CHECKING:
     from collections.abc import Collection, Sequence
@@ -147,49 +147,6 @@ def get_serial(ctx: click.Context) -> None:
     click.echo(f"Meter serial is: {response.serial}")
 
 
-REGISTERS_NAMES_LEN_MAX: Final[int] = max(
-    len(name) for name in constants.REGISTERS.values()
-)
-
-
-@attrs.define(kw_only=True)
-class RegisterOutput:
-    id_int: int
-    id_hex: str = attrs.field(init=False)
-    name: str = attrs.field(init=False)
-    unit_int: int
-    unit_hex: str = attrs.field(init=False)
-    unit_str: str = attrs.field(init=False)
-    value_float: float = attrs.field(init=False)
-    value_str: str = attrs.field(init=False)  # best: uses decimal.Decimal without loss
-    value_dec: decimal.Decimal
-
-    def __attrs_post_init__(self) -> None:
-        self.id_hex = f"0x{self.id_int:04X}"
-        self.unit_hex = f"0x{self.unit_int:02X}"
-        self.name = constants.REGISTERS.get(self.id_int, f"<unknown reg {self.id_int}>")
-        self.unit_str = constants.UNITS_NAMES.get(
-            self.unit_int, f"<unknown unit {self.unit_int}>"
-        )
-        self.value_float = float(self.value_dec)
-        self.value_str = str(self.value_dec)
-
-    @classmethod
-    def from_register_data(cls, reg: messages.RegisterData) -> Self:
-        value_dec = codec.FloatCodec.decode(reg.value)
-        return cls(
-            id_int=reg.id_,
-            unit_int=reg.unit,
-            value_dec=value_dec,
-        )
-
-    def to_pretty_line(self) -> str:
-        return (
-            f"{self.id_int!r:>4} → {self.name:<{REGISTERS_NAMES_LEN_MAX}} = "
-            f"{self.value_str} {self.unit_str}"
-        )
-
-
 def warn_registers_unknowns(
     registers: Collection[messages.RegisterData],
 ) -> None:
@@ -279,7 +236,7 @@ def get_register(
 
     warn_registers_unknowns(response.registers.values())
     outputs = (
-        RegisterOutput.from_register_data(reg) for reg in response.registers.values()
+        registers.RegisterOutput.from_register_data(reg) for reg in response.registers.values()
     )
 
     match output_format:
