+++ tests/test_registers.py
@@ -1,0 +1,45 @@
+# SPDX-FileCopyrightText: 2026 Jan Kundrát <jkt@jankundrat.com>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+import pytest
+import decimal
+
+from pykmp import messages, registers
+
+@pytest.mark.parametrize(
+    ("id_", "unit", "blob_with_size", "value_str", "value_dec", "unit_str"),
+    [
+        pytest.param(
+            1001,
+            51,
+            '04 00 00 00 00 00',
+            '0',
+            decimal.Decimal('0'),
+            'no unit (number)',
+        ),
+        pytest.param(
+            60,
+            8,
+            '04 43 00 05 c4 a6',
+            '378.022',
+            decimal.Decimal('378.022'),
+            'GJ',
+        ),
+    ]
+)
+def test_register_parsing(id_, unit, blob_with_size, value_str, value_dec, unit_str):
+    data = messages.RegisterData(id_=id_, unit=unit, value=bytes.fromhex(blob_with_size))
+    parsed = registers.RegisterOutput.from_register_data(data)
+    assert parsed.value_str == value_str
+    assert parsed.value_dec == value_dec
+    assert parsed.unit_str == unit_str
+
+
+def test_register_pretty_line():
+    data = messages.RegisterData(id_=1001, unit=51, value=bytes.fromhex('04 00 00 00 00 00'))
+    parsed = registers.RegisterOutput.from_register_data(data)
+    for part in ('1001', 'Fabrication No', ' 0 no unit (number)'):
+        assert part in parsed.to_pretty_line()
